Refactored 'Exceptions' from plural to singular
