+++ src/Exception/BadHttpRequestException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SkGovernmentParser\Exceptions;
+namespace SkGovernmentParser\Exception;
 
 
 class BadHttpRequestException extends \RuntimeException {}
