+++ src/Exception/EmptySearchResultException.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace SkGovernmentParser\Exceptions;
+namespace SkGovernmentParser\Exception;
 
 
 class EmptySearchResultException extends \RuntimeException {}
